+++ packages/plugin-dev-essentials/CHANGELOG.md
@@ -1,5 +1,280 @@
 # @business-as-code/plugin-dev-essentials
 
+## 0.0.0-latest-20230811071401
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811071401
+  - @business-as-code/core@0.0.0-latest-20230811071401
+  - @business-as-code/error@0.0.0-latest-20230811071401
+  - @business-as-code/fslib@0.0.0-latest-20230811071401
+
+## 0.0.0-latest-20230811071255
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811071255
+  - @business-as-code/core@0.0.0-latest-20230811071255
+  - @business-as-code/error@0.0.0-latest-20230811071255
+  - @business-as-code/fslib@0.0.0-latest-20230811071255
+
+## 0.0.0-latest-20230811065515
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811065515
+  - @business-as-code/core@0.0.0-latest-20230811065515
+  - @business-as-code/error@0.0.0-latest-20230811065515
+  - @business-as-code/fslib@0.0.0-latest-20230811065515
+
+## 0.0.0-latest-20230811065333
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811065333
+  - @business-as-code/core@0.0.0-latest-20230811065333
+  - @business-as-code/error@0.0.0-latest-20230811065333
+  - @business-as-code/fslib@0.0.0-latest-20230811065333
+
+## 0.0.0-latest-20230811065238
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811065238
+  - @business-as-code/core@0.0.0-latest-20230811065238
+  - @business-as-code/error@0.0.0-latest-20230811065238
+  - @business-as-code/fslib@0.0.0-latest-20230811065238
+
+## 0.0.0-latest-20230811064344
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811064344
+  - @business-as-code/core@0.0.0-latest-20230811064344
+  - @business-as-code/error@0.0.0-latest-20230811064344
+  - @business-as-code/fslib@0.0.0-latest-20230811064344
+
+## 0.0.0-latest-20230811064309
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811064309
+  - @business-as-code/core@0.0.0-latest-20230811064309
+  - @business-as-code/error@0.0.0-latest-20230811064309
+  - @business-as-code/fslib@0.0.0-latest-20230811064309
+
+## 0.0.0-latest-20230811063804
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811063804
+  - @business-as-code/core@0.0.0-latest-20230811063804
+  - @business-as-code/error@0.0.0-latest-20230811063804
+  - @business-as-code/fslib@0.0.0-latest-20230811063804
+
+## 0.0.0-latest-20230811063739
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811063739
+  - @business-as-code/core@0.0.0-latest-20230811063739
+  - @business-as-code/error@0.0.0-latest-20230811063739
+  - @business-as-code/fslib@0.0.0-latest-20230811063739
+
+## 0.0.0-latest-20230810180146
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810180146
+  - @business-as-code/core@0.0.0-latest-20230810180146
+  - @business-as-code/error@0.0.0-latest-20230810180146
+  - @business-as-code/fslib@0.0.0-latest-20230810180146
+
+## 0.0.0-latest-20230810175911
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810175911
+  - @business-as-code/core@0.0.0-latest-20230810175911
+  - @business-as-code/error@0.0.0-latest-20230810175911
+  - @business-as-code/fslib@0.0.0-latest-20230810175911
+
+## 0.0.0-latest-20230810175810
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810175810
+  - @business-as-code/core@0.0.0-latest-20230810175810
+  - @business-as-code/error@0.0.0-latest-20230810175810
+  - @business-as-code/fslib@0.0.0-latest-20230810175810
+
+## 0.0.0-latest-20230810175721
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810175721
+  - @business-as-code/core@0.0.0-latest-20230810175721
+  - @business-as-code/error@0.0.0-latest-20230810175721
+  - @business-as-code/fslib@0.0.0-latest-20230810175721
+
+## 0.0.0-latest-20230810174829
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810174829
+  - @business-as-code/core@0.0.0-latest-20230810174829
+  - @business-as-code/error@0.0.0-latest-20230810174829
+  - @business-as-code/fslib@0.0.0-latest-20230810174829
+
+## 0.0.0-latest-20230810174807
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810174807
+  - @business-as-code/core@0.0.0-latest-20230810174807
+  - @business-as-code/error@0.0.0-latest-20230810174807
+  - @business-as-code/fslib@0.0.0-latest-20230810174807
+
+## 0.0.0-latest-20230810174747
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810174747
+  - @business-as-code/core@0.0.0-latest-20230810174747
+  - @business-as-code/error@0.0.0-latest-20230810174747
+  - @business-as-code/fslib@0.0.0-latest-20230810174747
+
+## 0.0.0-latest-20230810173737
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810173737
+  - @business-as-code/core@0.0.0-latest-20230810173737
+  - @business-as-code/error@0.0.0-latest-20230810173737
+  - @business-as-code/fslib@0.0.0-latest-20230810173737
+
+## 0.0.0-latest-20230810172021
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810172021
+  - @business-as-code/core@0.0.0-latest-20230810172021
+  - @business-as-code/error@0.0.0-latest-20230810172021
+  - @business-as-code/fslib@0.0.0-latest-20230810172021
+
+## 0.0.0-latest-20230810171959
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171959
+  - @business-as-code/core@0.0.0-latest-20230810171959
+  - @business-as-code/error@0.0.0-latest-20230810171959
+  - @business-as-code/fslib@0.0.0-latest-20230810171959
+
+## 0.0.0-latest-20230810171910
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171910
+  - @business-as-code/core@0.0.0-latest-20230810171910
+  - @business-as-code/error@0.0.0-latest-20230810171910
+  - @business-as-code/fslib@0.0.0-latest-20230810171910
+
+## 0.0.0-latest-20230810171831
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171831
+  - @business-as-code/core@0.0.0-latest-20230810171831
+  - @business-as-code/error@0.0.0-latest-20230810171831
+  - @business-as-code/fslib@0.0.0-latest-20230810171831
+
+## 0.0.0-latest-20230810171226
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171226
+  - @business-as-code/core@0.0.0-latest-20230810171226
+  - @business-as-code/error@0.0.0-latest-20230810171226
+  - @business-as-code/fslib@0.0.0-latest-20230810171226
+
+## 0.0.0-latest-20230810171123
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171123
+  - @business-as-code/core@0.0.0-latest-20230810171123
+  - @business-as-code/error@0.0.0-latest-20230810171123
+  - @business-as-code/fslib@0.0.0-latest-20230810171123
+
+## 0.0.0-latest-20230810171057
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230810171057
+  - @business-as-code/core@0.0.0-latest-20230810171057
+  - @business-as-code/error@0.0.0-latest-20230810171057
+  - @business-as-code/fslib@0.0.0-latest-20230810171057
+
+## 0.0.0-bollards-20230810165657
+
+### Patch Changes
+
+- local snapshot
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230810165657
+  - @business-as-code/core@0.0.0-bollards-20230810165657
+  - @business-as-code/error@0.0.0-bollards-20230810165657
+  - @business-as-code/fslib@0.0.0-bollards-20230810165657
+
 ## 0.0.0-bollards-20230810141915
 
 ### Patch Changes